minui: drm: Delay enabling CRTC until we have rendered a frame
When drm_blank was called to re-enable the display, it set the
display to the current buffer. However, at this point the frame
was not rendered yet, which means you could slowly see it
rendering from top to bottom, instead of the full frame directly.

One option to solve this is to re-enable the display with the frame
we have rendered last, and wait until we can flip to next frame
(when it is ready). Especially for TWRP this looks weird though:
for a short time it displays the old, unlocked screen before jumping
to the screen look after a second.

Instead, we delay enabling the display until we have a frame ready
to be rendered. That means that the display will react slightly
slower after pressing the power button, but at least the correct
screen is shown.

Change-Id: Iff6673076d585535f889ef867b89b1627e7fc1b8