minui: drm: Render in software memory buffer
As for the other graphics implementations, rendering is very slow
when rendering directly to the mapped memory. Rendering is much
faster if we first use an extra buffer and then use memcpy to copy
it to the actual surface.

Another advantage is that we no longer have to wait for the page
flip to complete before we can start rendering. The old surface
buffer isn't modified until we are ready to schedule the next
surface.

Change-Id: I5d8d8c26fc3d3bb704f7597ffc85a456ebd2cd3f